Drop Python < 3.6
Drop py27/py35
Begin shifting to pytest
Shift to testing tox externally, rather than in-process

This will impact the coverage % in a decently significant way, as the
calls into the actual hooks will be done in a separate process. I don't
know if it is possible to grab that info with coverage, but I don't know
a simple way.