Drop support for old tox versions
In order to simplify maintenance of the project, we will support
only recent versions of tox and expose the minimal version in our
package metadata.

Since tox added support for bootstrapping itself requiring a minimal
version is no longer a problem.

This also fixes bug related to rename of whitelist, which prevented
used of tox-ansible by users using the new name.